warnings fixed

+++ src/eval.c
@@ -3,13 +3,13 @@ void operate(char* result, size_t len, char op, double num1, double num2)
     switch (op)
     {
         case '^':
-            snprintf(result, len, RSLT_ACCURACY_FORM, num1 ^ num2);
+            snprintf(result, len, RSLT_ACCURACY_FORM, (double) ((long) num1 ^ (long) num2));
             break;
         case '/':
             snprintf(result, len, RSLT_ACCURACY_FORM, num1 / num2);
             break;
         case '%':
-            snprintf(result, len, RSLT_ACCURACY_FORM, num1 % num2);
+            snprintf(result, len, RSLT_ACCURACY_FORM, (double) ((long) num1 % (long) num2));
             break;
         case '*':
             snprintf(result, len, RSLT_ACCURACY_FORM, num1 * num2);
